fix reward bag

+++ src/items/item.cpp
@@ -479,8 +479,8 @@ Attr_ReadValue Item::readAttr(AttrTypes_t attr, PropStream& propStream)
 		}
 
 		case ATTR_WRITTENDATE: {
-			uint32_t writtenDate;
-			if (!propStream.read<uint32_t>(writtenDate)) {
+			uint64_t writtenDate;
+			if (!propStream.read<uint64_t>(writtenDate)) {
 				return ATTR_READ_ERROR;
 			}
 
@@ -829,7 +829,7 @@ void Item::serializeAttr(PropWriteStream& propWriteStream) const
 	const time_t writtenDate = getDate();
 	if (writtenDate != 0) {
 		propWriteStream.write<uint8_t>(ATTR_WRITTENDATE);
-		propWriteStream.write<uint32_t>(writtenDate);
+		propWriteStream.write<uint64_t>(writtenDate);
 	}
 
 	const std::string& writer = getWriter();
