+++ src/items/item.h
@@ -51,7 +51,7 @@ class ItemAttributes
 			return getStrAttr(ITEM_ATTRIBUTE_TEXT);
 		}
 
-		void setDate(int32_t n) {
+		void setDate(int64_t n) {
 			setIntAttr(ITEM_ATTRIBUTE_DATE, n);
 		}
 		void resetDate() {
@@ -654,7 +654,7 @@ class Item : virtual public Thing
 			return getStrAttr(ITEM_ATTRIBUTE_TEXT);
 		}
 
-		void setDate(int32_t n) {
+		void setDate(int64_t n) {
 			setIntAttr(ITEM_ATTRIBUTE_DATE, n);
 		}
 		void resetDate() {
@@ -1052,7 +1052,7 @@ class Item : virtual public Thing
 		void addImbuement(uint8_t slot, uint16_t imbuementId, int32_t duration);
 		/**
 		 * @brief Decay imbuement time duration, only use this for decay the imbuement time
-		 * 
+		 *
 		 * @param slot Slot id to decay
 		 * @param imbuementId Imbuement id to decay
 		 * @param duration New duration
@@ -1144,7 +1144,7 @@ class Item : virtual public Thing
 
 		bool loadedFromMap = false;
 		bool isLootTrackeable = false;
-	
+
 	private:
 		void setImbuement(uint8_t slot, uint16_t imbuementId, int32_t duration);
 		//Don't add variables here, use the ItemAttribute class.
